Update patch set 15
Patch Set 15:

I believe this was marked WIP (and my terminal history says so) because I was just backuping these changes so I can transfer them easily to my work computer at school.

I wanted to discuss some parts of it with you on the call today. I will go through your notes when I get to school. But this change is not a priority. It "works" but it is not ready.

Patch-set: 15