Use patched systemd for build
Today, systemd does not export any API for publishing the LLDP data
obtained via systemd-networkd. Originally, we wanted to use D-Bus, but
based on a discussion with system upstream, it was decided to publish
the parsing code into system'd C API instead. This now means that we
require patched systemd not just at runtime, but for the build itself as
well.

This patch configures Zuul to provide a properly patched systemd source
tree, which should then -- hopefully -- be picked by CMake as usual.

Make sure that we do not change the build env (too) much, because
apparently this particular version of systemd cannot be built with, say,
UBSAN just via setting the CFLAGS and CXXFLAGS and LDFLAGS -- hence the
`env -u` dance. Also, praise to the dnf builddep plugin for not making
me enumerate these 90 packages that get installed by hand, yay.

Bug: https://github.com/systemd/systemd/pull/16744
Change-Id: I5a787f277b1d7e2bcc0226529917906405dbca9a