tests: Do not assume bash is at /bin/bash
Yay, a NixOS change. See, in a `nix-shell` bash can indeed be available,
but it cannot be made available as /bin/bash. Interestingly enough,
POSIX doesn't even mandate that the classic Bourne shell be at /bin/sh,
but there is a mandate for a standard /usr/bin/env. So let's make my
life easier on this crazy Nix.

Change-Id: I00877d8a6a5083cba6d7e1edd651e27392af5c52