Fix creating if-name lists in sysrepo callback
While working on another bug [1] we found out that the if-name list creates one entry
for all neighbors. It should create one entry per neighbor.

This was left unnoticed due to insufficient testing. We only tried to
push a single neighbor into sysrepo so the bug was not triggered. This
is now changed.

[1] https://github.com/CESNET/lldp-systemd-networkd-sysrepo/issues/1

Change-Id: I0c501242b3b57a2973f14f83710ff1d6197e2473